chore: fix clippy warnings and enforce -D warnings in pre-commit
- Add crate-level allows for dead_code, type_complexity, too_many_arguments
- Fix map_or -> is_some_and conversions
- Fix to_local self convention for Copy types
- Simplify weather_icon placeholder function
- Remove unused imports
- Update pre-commit hook to fail on warnings

+++ src/render/graphics.rs
@@ -62,7 +62,7 @@ impl Graphics {
         attr_string.replace_str(&cf_string, CFRange::init(0, 0));
 
         // Use CFString length (UTF-16 code units), not byte length
-        let range = CFRange::init(0, cf_string.char_len() as isize);
+        let range = CFRange::init(0, cf_string.char_len());
 
         // Set font attribute using the kCTFontAttributeName constant
         unsafe {
@@ -135,7 +135,7 @@ impl Graphics {
         attr_string.replace_str(&cf_string, CFRange::init(0, 0));
 
         // Use CFString length (UTF-16 code units), not byte length
-        let range = CFRange::init(0, cf_string.char_len() as isize);
+        let range = CFRange::init(0, cf_string.char_len());
 
         unsafe {
             attr_string.set_attribute(range, kCTFontAttributeName, &self.font);
